# LabelSet.String: restore faster sort call